moving database handling logic to repository
