--- app/src/main/java/android/support/v17/leanback/supportleanbackshowcase/app/room/db/DatabaseInitUtil.java
@@ -1,144 +1,0 @@
-/*
- * Copyright (C) 2014 The Android Open Source Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License. You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under the License
- * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
- * or implied. See the License for the specific language governing permissions and limitations under
- * the License.
- */
-
-package android.support.v17.leanback.supportleanbackshowcase.app.room.db;
-
-import android.os.AsyncTask;
-import android.support.v17.leanback.supportleanbackshowcase.R;
-import android.support.v17.leanback.supportleanbackshowcase.app.room.SampleApplication;
-import android.support.v17.leanback.supportleanbackshowcase.app.room.api.VideoDownloadingService;
-import android.support.v17.leanback.supportleanbackshowcase.app.room.api.VideosWithGoogleTag;
-import android.support.v17.leanback.supportleanbackshowcase.app.room.config.AppConfiguration;
-import android.support.v17.leanback.supportleanbackshowcase.app.room.db.entity.CategoryEntity;
-import android.support.v17.leanback.supportleanbackshowcase.app.room.db.entity.VideoEntity;
-import android.support.v17.leanback.supportleanbackshowcase.utils.Utils;
-import android.util.Log;
-import android.widget.PopupMenu;
-
-import com.google.gson.Gson;
-
-import java.io.IOException;
-
-import retrofit2.Call;
-import retrofit2.Callback;
-import retrofit2.Response;
-import retrofit2.Retrofit;
-import retrofit2.converter.gson.GsonConverterFactory;
-
-public class DatabaseInitUtil {
-    // For debugging purpose
-    private static final boolean DEBUG = false;
-    private static final String TAG = "VideoDbBuilder";
-
-    public void initializeDb(AppDatabase db, String url) throws IOException {
-
-        // json data
-        String json;
-
-        if (AppConfiguration.IS_DEBUGGING_VERSION) {
-
-            // when use debugging version, we won't fetch data from network but using local
-            // json file (only contain 4 video entities in 2 categories.)
-            json = Utils.inputStreamToString(SampleApplication
-                    .getInstance()
-                    .getApplicationContext()
-                    .getResources()
-                    .openRawResource(R.raw.live_movie_debug));
-            Gson gson = new Gson();
-            VideosWithGoogleTag videosWithGoogleTag = gson.fromJson(json,
-                VideosWithGoogleTag.class);
-            populateDatabase(videosWithGoogleTag,db);
-        } else {
-            buildDatabase(db, url);
-        }
-    }
-
-    /**
-     * Takes the contents of a JSON object and populates the database
-     *
-     * @param db Room database.
-     */
-    public static void buildDatabase(final AppDatabase db, String url) throws IOException {
-        Retrofit retrofit = new Retrofit
-                .Builder()
-                .baseUrl(url)
-                .addConverterFactory(GsonConverterFactory.create())
-                .build();
-        VideoDownloadingService service = retrofit.create(VideoDownloadingService.class);
-
-        Call<VideosWithGoogleTag> videosWithGoogleTagCall = service.getVideosList();
-        videosWithGoogleTagCall.enqueue(new Callback<VideosWithGoogleTag>() {
-            @Override
-            public void onResponse(Call<VideosWithGoogleTag> call, Response<VideosWithGoogleTag> response) {
-                VideosWithGoogleTag videosWithGoogleTag = response.body();
-                if (videosWithGoogleTag == null) {
-                    Log.d(TAG, "onResponse: result is null");
-                    return;
-                }
-                populateDatabase(videosWithGoogleTag, db);
-            }
-
-            @Override
-            public void onFailure(Call<VideosWithGoogleTag> call, Throwable t) {
-                Log.d(TAG, "Fail to download the content");
-            }
-        });
-
-    }
-
-    private static void populateDatabase(VideosWithGoogleTag videosWithGoogleTag, final AppDatabase db) {
-        for (final VideosWithGoogleTag.VideosGroupByCategory videosGroupByCategory :
-                videosWithGoogleTag.getAllResources()) {
-
-            // create category table
-            final CategoryEntity categoryEntity = new CategoryEntity();
-            categoryEntity.setCategoryName(videosGroupByCategory.getCategory());
-
-            // create video table with customization
-            postProcessing(videosGroupByCategory);
-
-            new AsyncTask<Void, Void, Void>() {
-                @Override
-                protected Void doInBackground(Void... voids) {
-                    try {
-                        db.beginTransaction();
-                        db.categoryDao().insertCategory(categoryEntity);
-                        db.videoDao().insertAllVideos(videosGroupByCategory.getVideos());
-                        db.setTransactionSuccessful();
-                    } finally {
-                        db.endTransaction();
-                    }
-                    return null;
-                }
-            }.execute();
-
-        }
-    }
-
-    /**
-     * Helper function to make some customization on raw data
-     */
-    private static void postProcessing(VideosWithGoogleTag.VideosGroupByCategory videosGroupByCategory) {
-        for (VideoEntity each : videosGroupByCategory.getVideos()) {
-            each.setCategory(videosGroupByCategory.getCategory());
-            each.setVideoLocalStorageUrl("");
-            each.setVideoBgImageLocalStorageUrl("");
-            each.setVideoCardImageLocalStorageUrl("");
-            each.setVideoUrl(each.getVideoUrls().get(0));
-            each.setRented(false);
-            each.setStatus("");
-            each.setTrailerVideoUrl("https://storage.googleapis.com/android-tv/Sample%20videos/Google%2B/Google%2B_%20Say%20more%20with%20Hangouts.mp4");
-        }
-    }
-}
