+++ app/src/main/java/android/support/v17/leanback/supportleanbackshowcase/app/room/db/repo/VideosRepository.java
@@ -18,20 +18,50 @@
 package android.support.v17.leanback.supportleanbackshowcase.app.room.db.repo;
 
 import android.arch.lifecycle.LiveData;
+import android.arch.persistence.room.Room;
+import android.os.AsyncTask;
+import android.support.annotation.WorkerThread;
+import android.support.v17.leanback.supportleanbackshowcase.R;
 import android.support.v17.leanback.supportleanbackshowcase.app.room.SampleApplication;
+import android.support.v17.leanback.supportleanbackshowcase.app.room.api.VideoDownloadingService;
+import android.support.v17.leanback.supportleanbackshowcase.app.room.api.VideosWithGoogleTag;
+import android.support.v17.leanback.supportleanbackshowcase.app.room.config.AppConfiguration;
 import android.support.v17.leanback.supportleanbackshowcase.app.room.db.AppDatabase;
-import android.support.v17.leanback.supportleanbackshowcase.app.room.db.DatabaseHelper;
 import android.support.v17.leanback.supportleanbackshowcase.app.room.db.dao.CategoryDao;
 import android.support.v17.leanback.supportleanbackshowcase.app.room.db.dao.VideoDao;
 import android.support.v17.leanback.supportleanbackshowcase.app.room.db.entity.CategoryEntity;
 import android.support.v17.leanback.supportleanbackshowcase.app.room.db.entity.VideoEntity;
+import android.support.v17.leanback.supportleanbackshowcase.utils.Utils;
+import android.util.Log;
 
+import com.google.gson.Gson;
+
+import java.io.IOException;
 import java.util.HashMap;
 import java.util.List;
 
+import retrofit2.Call;
+import retrofit2.Callback;
+import retrofit2.Response;
+import retrofit2.Retrofit;
+import retrofit2.converter.gson.GsonConverterFactory;
+
 public class VideosRepository {
+
+    // For debugging purpose
+    private static final boolean DEBUG = false;
+    private static final String TAG = "VideosRepository";
+
+    // meta data type constant
+    private static final String RENTED = "rented";
+    private static final String STATUS = "status";
+    private static final String CARD = "card";
+    private static final String BACKGROUND = "background";
+    private static final String VIDEO = "video";
+
     private static VideosRepository sVideosRepository;
 
+    private AppDatabase mDb;
     private VideoDao mVideoDao;
     private CategoryDao mCategoryDao;
 
@@ -43,13 +73,6 @@ public class VideosRepository {
             sVideosRepository = new VideosRepository();
         }
         return sVideosRepository;
-    }
-
-    private VideosRepository() {
-        AppDatabase appDatabase = DatabaseHelper.getInstance().getDatabase(SampleApplication.getInstance());
-        mVideoDao = appDatabase.videoDao();
-        mCategoryDao = appDatabase.categoryDao();
-        mRepositoryCache = new HashMap<>();
     }
 
     public LiveData<List<VideoEntity>> getVideosInSameCategoryLiveData(String category) {
@@ -71,5 +94,168 @@ public class VideosRepository {
 
     public LiveData<VideoEntity> getVideoById(Long id) {
         return mVideoDao.loadVideoById(id);
+    }
+
+
+
+    /**
+     * Helper function to access the database and update the video information in the database.
+     *
+     * @param video    video entity
+     * @param category which fields to update
+     * @param value    updated value
+     */
+    @WorkerThread
+    public synchronized void updateDatabase(VideoEntity video, String category, String value) {
+        try {
+            mDb.beginTransaction();
+            switch (category) {
+                case VIDEO:
+                    video.setVideoLocalStorageUrl(value);
+                    break;
+                case BACKGROUND:
+                    video.setVideoBgImageLocalStorageUrl(value);
+                    break;
+                case CARD:
+                    video.setVideoCardImageLocalStorageUrl(value);
+                    break;
+                case STATUS:
+                    video.setStatus(value);
+                    break;
+                case RENTED:
+                    video.setRented(true);
+                    break;
+            }
+            mDb.videoDao().updateVideo(video);
+            mDb.setTransactionSuccessful();
+        } finally {
+            mDb.endTransaction();
+        }
+    }
+
+    public void reCreateDatabaseIfNecessary() {
+        createAndPopulateDatabase();
+    }
+
+    private VideosRepository() {
+        createAndPopulateDatabase();
+        mVideoDao = mDb.videoDao();
+        mCategoryDao = mDb.categoryDao();
+        mRepositoryCache = new HashMap<>();
+    }
+
+    private void createAndPopulateDatabase() {
+        mDb = Room.databaseBuilder(SampleApplication.getInstance(),
+                AppDatabase.class, AppDatabase.DATABASE_NAME).build();
+
+        // insert contents into database
+        try {
+            String url =
+                    "https://storage.googleapis.com/android-tv/";
+            initializeDb(mDb, url);
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+    }
+
+    private void initializeDb(AppDatabase db, String url) throws IOException {
+
+        // json data
+        String json;
+
+        if (AppConfiguration.IS_DEBUGGING_VERSION) {
+
+            // when use debugging version, we won't fetch data from network but using local
+            // json file (only contain 4 video entities in 2 categories.)
+            json = Utils.inputStreamToString(SampleApplication
+                    .getInstance()
+                    .getApplicationContext()
+                    .getResources()
+                    .openRawResource(R.raw.live_movie_debug));
+            Gson gson = new Gson();
+            VideosWithGoogleTag videosWithGoogleTag = gson.fromJson(json,
+                    VideosWithGoogleTag.class);
+            populateDatabase(videosWithGoogleTag,db);
+        } else {
+            buildDatabase(db, url);
+        }
+    }
+
+    /**
+     * Takes the contents of a JSON object and populates the database
+     *
+     * @param db Room database.
+     */
+    private static void buildDatabase(final AppDatabase db, String url) throws IOException {
+        Retrofit retrofit = new Retrofit
+                .Builder()
+                .baseUrl(url)
+                .addConverterFactory(GsonConverterFactory.create())
+                .build();
+        VideoDownloadingService service = retrofit.create(VideoDownloadingService.class);
+
+        Call<VideosWithGoogleTag> videosWithGoogleTagCall = service.getVideosList();
+        videosWithGoogleTagCall.enqueue(new Callback<VideosWithGoogleTag>() {
+            @Override
+            public void onResponse(Call<VideosWithGoogleTag> call, Response<VideosWithGoogleTag> response) {
+                VideosWithGoogleTag videosWithGoogleTag = response.body();
+                if (videosWithGoogleTag == null) {
+                    Log.d(TAG, "onResponse: result is null");
+                    return;
+                }
+                populateDatabase(videosWithGoogleTag, db);
+            }
+
+            @Override
+            public void onFailure(Call<VideosWithGoogleTag> call, Throwable t) {
+                Log.d(TAG, "Fail to download the content");
+            }
+        });
+
+    }
+
+    private static void populateDatabase(VideosWithGoogleTag videosWithGoogleTag, final AppDatabase db) {
+        for (final VideosWithGoogleTag.VideosGroupByCategory videosGroupByCategory :
+                videosWithGoogleTag.getAllResources()) {
+
+            // create category table
+            final CategoryEntity categoryEntity = new CategoryEntity();
+            categoryEntity.setCategoryName(videosGroupByCategory.getCategory());
+
+            // create video table with customization
+            postProcessing(videosGroupByCategory);
+
+            new AsyncTask<Void, Void, Void>() {
+                @Override
+                protected Void doInBackground(Void... voids) {
+                    try {
+                        db.beginTransaction();
+                        db.categoryDao().insertCategory(categoryEntity);
+                        db.videoDao().insertAllVideos(videosGroupByCategory.getVideos());
+                        db.setTransactionSuccessful();
+                    } finally {
+                        db.endTransaction();
+                    }
+                    return null;
+                }
+            }.execute();
+
+        }
+    }
+
+    /**
+     * Helper function to make some customization on raw data
+     */
+    private static void postProcessing(VideosWithGoogleTag.VideosGroupByCategory videosGroupByCategory) {
+        for (VideoEntity each : videosGroupByCategory.getVideos()) {
+            each.setCategory(videosGroupByCategory.getCategory());
+            each.setVideoLocalStorageUrl("");
+            each.setVideoBgImageLocalStorageUrl("");
+            each.setVideoCardImageLocalStorageUrl("");
+            each.setVideoUrl(each.getVideoUrls().get(0));
+            each.setRented(false);
+            each.setStatus("");
+            each.setTrailerVideoUrl("https://storage.googleapis.com/android-tv/Sample%20videos/Google%2B/Google%2B_%20Say%20more%20with%20Hangouts.mp4");
+        }
     }
 }
